reacjilator-python3/requirements.txt
google-cloud-translate==1.4.0
slackclient==2.0.0
slackeventsapi==2.1.0
